requirements.txt
langchain
langchain-openai
PyPDF2
langgraph
streamlit
langfuse
serpapi
google-search-results
python-dotenv